# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Together API",
    "description": "Private, non-federated Discord alternative for small gaming communities",
    "contact": {
      "name": "Together Contributors"
    },
    "license": {
      "name": "PolyForm Noncommercial 1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/channels/{channel_id}/ack": {
      "post": {
        "tags": [
          "ReadStates"
        ],
        "summary": "POST /channels/:channel_id/ack — mark a server channel as read.",
        "description": "Upserts the user's read-state to `NOW()`.  Called by the client whenever\nthe user views a channel (tab focus, navigation).",
        "operationId": "ack_channel",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Channel marked as read"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/channels/{channel_id}/messages/{message_id}/reactions": {
      "get": {
        "tags": [
          "Reactions"
        ],
        "summary": "GET /channels/:channel_id/messages/:message_id/reactions",
        "description": "List aggregated reaction counts for a message, with a `me` flag indicating\nwhether the authenticated user has added each reaction.",
        "operationId": "list_reactions",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reaction counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReactionCount"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/channels/{channel_id}/messages/{message_id}/reactions/{emoji}": {
      "put": {
        "tags": [
          "Reactions"
        ],
        "summary": "PUT /channels/:channel_id/messages/:message_id/reactions/:emoji",
        "description": "Add an emoji reaction to a message.  Idempotent — adding the same emoji\ntwice is not an error (the duplicate is silently ignored).",
        "operationId": "add_reaction",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "emoji",
            "in": "path",
            "description": "Emoji string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Reaction added"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Reactions"
        ],
        "summary": "DELETE /channels/:channel_id/messages/:message_id/reactions/:emoji",
        "description": "Remove the authenticated user's reaction from a message.\nReturns 404 if the message or reaction does not exist.",
        "operationId": "remove_reaction",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "emoji",
            "in": "path",
            "description": "Emoji string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Reaction removed"
          },
          "404": {
            "description": "Reaction not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/dm-channels/{id}/ack": {
      "post": {
        "tags": [
          "ReadStates"
        ],
        "summary": "POST /dm-channels/:channel_id/ack — mark a DM channel as read.",
        "operationId": "ack_dm_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DM channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "DM channel marked as read"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReactionCount": {
        "type": "object",
        "description": "Aggregated reaction count for a single emoji on a message.",
        "required": [
          "emoji",
          "count",
          "me"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "emoji": {
            "type": "string"
          },
          "me": {
            "type": "boolean",
            "description": "Whether the requesting user has added this reaction at the time of the\nquery. Clients must apply `REACTION_ADD`/`REACTION_REMOVE` WebSocket\ndelta events to keep this accurate after the initial fetch."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}